+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [0.5.0](https://github.com/dldevinc/paper-uploads/tree/v0.5.0) - 2021-04-03
+- Removed default permissions from all models.
 - Add Python 3.9 support
 
 ## [0.4.5](https://github.com/dldevinc/paper-uploads/tree/v0.4.5) - 2020-10-26